requirements.txt
terminaltables==3.1.10
requests==2.31.0
python-dotenv==1.0.4